chore(refactor): changed the project structure
Related work item: GH-6
