+++ index.html
@@ -17,7 +17,7 @@
     <div id="root"></div>
     <script
       type="module"
-      src="/src/main.tsx"
+      src="/src/pages/main/main.tsx"
     ></script>
   </body>
 </html>
