--- src/App.tsx
@@ -1,8 +1,0 @@
-import { Footer } from '../components/Footer';
-export function App() {
-  return (
-    <div className='App'>
-      <Footer></Footer>
-    </div>
-  );
-}
